move maps to own package mapsutils
